Push details of directive name lookup into the cost spec so abstraction takes schema instead
